Try to enter the first week by hand to see what we are up against
requirements.txt
- some stuff is not needed (well, a lot of stuff is not needed) and we
  will get to that in time.

added migrations.  I think we will torch most of the crap I am doing
now, but it is worth just posting it for future reference.

workout/models.py
- added autoslugfield.  Seems like this is an easy solution and so far
  there has not been any conflicts.  Only time will tell if this is a
  good idea.
- add Autoslugfield and ManyToManyField relationships to clean the
  code up a bit.
- sequence field really does not need a help text, so I torched it.
- added __str__ to show (maybe?) what the workout is.

flake8 with python 3.8 is waaaay stricter than 3.7 and earlier, so I
needed to put in comments to shut flake8 up.